Move user-facing types to src/api/
Extract MockClient, StubClient, StubBuilderFor, FetchPageStubBuilder,
FetchOneStubBuilder, AggregateStubBuilder, QueryStubBuilder, and
MockOsdkObjectOptions into dedicated files under src/api/. Stub builders
are grouped in StubBuilders.ts; MockClient, StubClient, and
MockOsdkObjectOptions each have their own file. Public API surface is
unchanged.
