+++ packages/functions-testing.experimental/src/api/MockClient.ts
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2026 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type {
+  CompileTimeMetadata,
+  ObjectOrInterfaceDefinition,
+  ObjectSet,
+  QueryDefinition,
+} from "@osdk/api";
+import type { Client } from "@osdk/client";
+import type { QueryStubBuilder, StubBuilderFor } from "./StubBuilders.js";
+import type { StubClient } from "./StubClient.js";
+
+type QueryReturnTypeFromDef<Q extends QueryDefinition> = ReturnType<
+  CompileTimeMetadata<Q>["signature"]
+> extends Promise<infer R> ? R : never;
+
+type QueryParamsFromDef<Q extends QueryDefinition> =
+  Parameters<CompileTimeMetadata<Q>["signature"]> extends [infer P] ? P
+    : undefined;
+
+export type StubPatternCallback<T> = (client: StubClient) => T;
+
+export type ObjectSetStubCallback<Q extends ObjectOrInterfaceDefinition, T> = (
+  os: ObjectSet<Q>,
+) => T;
+
+export interface MockClient extends Client {
+  when<T>(callback: StubPatternCallback<T>): StubBuilderFor<T>;
+  whenObjectSet<Q extends ObjectOrInterfaceDefinition, T>(
+    objectSet: ObjectSet<Q>,
+    callback: ObjectSetStubCallback<Q, T>,
+  ): StubBuilderFor<T>;
+  whenQuery<Q extends QueryDefinition>(
+    query: Q,
+    params?: QueryParamsFromDef<Q>,
+  ): QueryStubBuilder<QueryReturnTypeFromDef<Q>>;
+  clearStubs(): void;
+}
