+++ packages/functions-testing.experimental/src/api/StubClient.ts
@@ -14,18 +14,13 @@
  * limitations under the License.
  */
 
-export type {
-  AggregateStubBuilder,
-  FetchOneStubBuilder,
-  FetchPageStubBuilder,
-  MockClient,
-  MockOsdkObjectOptions,
-  QueryStubBuilder,
-  StubBuilderFor,
-  StubClient,
-} from "./api/index.js";
+import type {
+  InterfaceDefinition,
+  ObjectSet,
+  ObjectTypeDefinition,
+} from "@osdk/api";
 
-export { createMockAttachment } from "./mock/createMockAttachment.js";
-export { createMockClient } from "./mock/createMockClient.js";
-export { createMockObjectSet } from "./mock/createMockObjectSet.js";
-export { createMockOsdkObject } from "./mock/createMockOsdkObject.js";
+export type StubClient = {
+  <Q extends ObjectTypeDefinition>(o: Q): ObjectSet<Q>;
+  <Q extends InterfaceDefinition>(o: Q): ObjectSet<Q>;
+};
